refactor: 작가 리스트 조회 수정
