+++ src/main/java/com/nhnacademy/inkbridge/backend/repository/BookAuthorRepository.java
@@ -2,7 +2,6 @@ package com.nhnacademy.inkbridge.backend.repository;
 
 import com.nhnacademy.inkbridge.backend.entity.BookAuthor;
 import com.nhnacademy.inkbridge.backend.entity.BookAuthor.Pk;
-import com.nhnacademy.inkbridge.backend.repository.custom.BookAuthorRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 /**
@@ -11,10 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author minm063
  * @version 2024/02/15
  */
-public interface BookAuthorRepository extends JpaRepository<BookAuthor, Pk>,
-    BookAuthorRepositoryCustom {
-
-    BookAuthor findByPk_BookId(Long bookId);
+public interface BookAuthorRepository extends JpaRepository<BookAuthor, Pk> {
 
     void deleteAllByBook_BookId(Long bookId);
 }
